pyproject.toml
[project]
name = "test-task-2025-03-31"
version = "0.1.0"
description = ""
authors = [
    {name = "vktadm",email = "tarnavskayavictoria@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "environs (>=14.1.1,<15.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
